move chrome requests function to helper
